third_party/WebKit/LayoutTests/inspector/timeline/timeline-layout-autosize-expected.txt

Tests the Timeline API instrumentation of a Layout event with Text Autosizing enabled.

text
text
text
text
text
AutosizeText Properties:
{
    children : <object>
    data : {
        backendNodeId : <number>
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    type : "AutosizeText"
}
AutosizeText Properties:
{
    children : <object>
    data : {
        backendNodeId : <number>
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    type : "AutosizeText"
}
